app/event-service/requirements.txt
annotated-types==0.7.0
anyio==4.1.2.0
click==8.3.1
colorama==0.4.6
fastapi==0.127.0
h11==0.16.0
idna==3.11
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
sqlmodel==0.0.22
sqlalchemy==2.0.36
